KM451: Move functionality from github direct into githib service
