--- pkg/client/core/api/direct/doc.go
@@ -1,2 +1,0 @@
-// Package direct implements direct usage of the server side services
-package direct
