+++ pkg/client/core/service_github.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/oslokommune/okctl/pkg/api"
+	"github.com/oslokommune/okctl/pkg/keypair"
 
 	"github.com/oslokommune/okctl/pkg/github"
 
@@ -19,6 +20,10 @@ type githubService struct {
 	state client.GithubState
 }
 
+func githubDeployKeySecretName(org, repo string) string {
+	return fmt.Sprintf("github/deploykeys/%s/%s/privatekey", org, repo)
+}
+
 func (s *githubService) DeleteGithubRepository(_ context.Context, opts client.DeleteGithubRepositoryOpts) error {
 	fullName := fmt.Sprintf("%s/%s", opts.Organisation, opts.Name)
 
@@ -31,7 +36,7 @@ func (s *githubService) DeleteGithubRepository(_ context.Context, opts client.De
 		return err
 	}
 
-	err = s.api.DeleteRepositoryDeployKey(client.DeleteGithubDeployKeyOpts{
+	err = s.DeleteRepositoryDeployKey(client.DeleteGithubDeployKeyOpts{
 		ID:           opts.ID,
 		Organisation: opts.Organisation,
 		Repository:   opts.Name,
@@ -58,7 +63,7 @@ func (s *githubService) CreateGithubRepository(_ context.Context, opts client.Cr
 		return r, nil
 	}
 
-	key, err := s.api.CreateRepositoryDeployKey(client.CreateGithubDeployKeyOpts{
+	key, err := s.CreateRepositoryDeployKey(client.CreateGithubDeployKeyOpts{
 		ID:           opts.ID,
 		Organisation: opts.Organization,
 		Repository:   opts.Name,
@@ -85,8 +90,53 @@ func (s *githubService) CreateGithubRepository(_ context.Context, opts client.Cr
 	return repo, nil
 }
 
-func (s *githubService) ListReleases(owner, repo string) ([]*github.RepositoryRelease, error) {
-	return s.api.ListReleases(owner, repo)
+func (a *githubService) DeleteRepositoryDeployKey(opts client.DeleteGithubDeployKeyOpts) error {
+	err := a.parameterService.DeleteSecret(context.Background(), api.DeleteSecretOpts{
+		Name: githubDeployKeySecretName(opts.Organisation, opts.Repository),
+	})
+	if err != nil {
+		return err
+	}
+
+	return a.githubClient.DeleteDeployKey(opts.Organisation, opts.Repository, opts.Identifier)
+}
+
+func (a *githubService) ListReleases(owner, repo string) ([]*github.RepositoryRelease, error) {
+	return a.githubClient.ListReleases(owner, repo)
+}
+
+func (a *githubService) CreateRepositoryDeployKey(opts client.CreateGithubDeployKeyOpts) (*client.GithubDeployKey, error) {
+	key, err := keypair.New(keypair.DefaultRandReader(), keypair.DefaultBitSize).Generate()
+	if err != nil {
+		return nil, err
+	}
+
+	param, err := a.parameterService.CreateSecret(context.Background(), api.CreateSecretOpts{
+		ID:     opts.ID,
+		Name:   githubDeployKeySecretName(opts.Organisation, opts.Repository),
+		Secret: string(key.PrivateKey),
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	deployKey, err := a.githubClient.CreateDeployKey(opts.Organisation, opts.Repository, opts.Title, string(key.PublicKey))
+	if err != nil {
+		return nil, err
+	}
+
+	return &client.GithubDeployKey{
+		Organisation: opts.Organisation,
+		Repository:   opts.Repository,
+		Identifier:   deployKey.GetID(),
+		Title:        deployKey.GetTitle(),
+		PublicKey:    string(key.PublicKey),
+		PrivateKeySecret: &client.GithubSecret{
+			Name:    param.Name,
+			Path:    param.Path,
+			Version: param.Version,
+		},
+	}, nil
 }
 
 // NewGithubService returns an initialised service
